DishDetailScreen basic functionality and UI

+++ shared/src/commonMain/kotlin/com/example/havira/dish/data/local/DishMapper.kt
@@ -14,7 +14,7 @@ fun DishEntity.toDish(): Dish {
         title = title,
         desc = description,
         created = Instant.fromEpochMilliseconds(created).toLocalDateTime(TimeZone.currentSystemDefault()),
-        rating = rating.toInt(),
+        rating = rating.toFloat(),
         nofRatings = nof_ratings.toInt(),
         lastMade = last_made?.let { Instant.fromEpochMilliseconds(it).toLocalDateTime(TimeZone.currentSystemDefault()) }
     )
